// src/common/config/enum.d.ts
export declare enum ConfigKey {
    DB_TYPE = "DB_TYPE",
    DB_HOST = "DB_HOST",
    DB_PORT = "DB_PORT",
    DB_USER = "DB_USER",
    DB_PASSWORD = "DB_PASSWORD",
    DB_DATABASE = "DB_DATABASE",
    DB_SYNC = "DB_SYNC",
    DB_MIGRATIONS = "DB_MIGRATIONS",
    APP_PORT = "APP_PORT",
    APP_MODE = "APP_MODE",
    APP_BASE_URL = "APP_BASE_URL",
    JWT_TOKEN_SECRET = "JWT_TOKEN_SECRET",
    JWT_TOKEN_EXPIRE_IN = "JWT_TOKEN_EXPIRE_IN",
    JWT_REFRESH_TOKEN_SECRET = "JWT_REFRESH_TOKEN_SECRET",
    JWT_REFRESH_TOKEN_EXPIRE_IN = "JWT_REFRESH_TOKEN_EXPIRE_IN"
}
export declare const configMinimalKeys: ConfigKey[];